use next/image, rename Header to HeroHeader
