+++ src/app/page.tsx
@@ -4,7 +4,7 @@ import HeaderFeature from '@/components/home/header-feature'
 import { StoryCards } from '@/components/home/story-cards/StoryCards'
 import { MainContainer } from '@/components/shared/containers/MainContainer'
 import { Footer } from '@/components/shared/footer'
-import { Header } from '@/components/shared/header'
+import { HeroHeader } from '@/components/shared/hero-header/HeroHeader'
 import { LinkButton } from '@/components/shared/link-button'
 import { Navbar } from '@/components/shared/navbar'
 import { SemanticHeader } from '@/components/shared/SemanticHeader'
@@ -16,13 +16,13 @@ export default function Home() {
   return (
     <>
       <SemanticHeader>
-        <Header hasBorder>
+        <HeroHeader hasBorder>
           <Navbar>
             <SiteLogo />
             <LinkButton href={SIGN_IN}>Sign In</LinkButton>
           </Navbar>
           <HeaderFeature />
-        </Header>
+        </HeroHeader>
       </SemanticHeader>
 
       <MainContainer>
