+++ src/components/shared/hero-header/styled.ts
@@ -1,24 +1,16 @@
+import Image from 'next/image'
 import styled from 'styled-components'
 
 import { cardBorderBottom } from '@/constants/theme'
 
 type HeaderProps = {
   $hasBorder?: boolean
-  $noBgOnMobile?: boolean
 }
 
 export const HeaderContainer = styled.div<HeaderProps>`
+  position: relative;
   width: 100%;
   border-bottom: ${({ $hasBorder }) => ($hasBorder ? cardBorderBottom : '')};
-  background-image: ${({ $noBgOnMobile }) =>
-    $noBgOnMobile ? `url('')` : `url('./images/misc/videodrome.jpg')`};
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-position: center;
-
-  @media (min-width: 740px) {
-    background-image: url('./images/misc/videodrome.jpg');
-  }
 `
 export const ImageGradient = styled.div`
   background: rgba(0, 0, 0, 0.4);
@@ -28,4 +20,20 @@ export const ImageGradient = styled.div`
     rgba(0, 0, 0, 0) 60%,
     rgba(0, 0, 0, 0.8) 100%
   );
+`
+
+type BackgroundImageProps = {
+  $noBgOnMobile?: boolean
+}
+
+export const BackgroundImage = styled(Image)<BackgroundImageProps>`
+  z-index: -1;
+  object-fit: cover;
+  position: center;
+
+  display: ${({ $noBgOnMobile }) => ($noBgOnMobile ? 'none' : 'block')};
+
+  @media (min-width: 740px) {
+    display: block;
+  }
 `
